added a cast to avoid an integer overflow on 32bit machines.
2007-11-17  Sven Neumann  <sven@gimp.org>

	* app/base/base-utils.c (get_physical_memory_size): added a cast
	to avoid an integer overflow on 32bit machines.


svn path=/trunk/; revision=24182

+++ app/base/base-utils.c
@@ -61,7 +61,7 @@ get_physical_memory_size (void)
 {
 #ifdef G_OS_UNIX
 #if defined(HAVE_UNISTD_H) && defined(_SC_PHYS_PAGES) && defined (_SC_PAGE_SIZE)
-  return sysconf (_SC_PHYS_PAGES) * sysconf (_SC_PAGE_SIZE);
+  return (guint64) sysconf (_SC_PHYS_PAGES) * sysconf (_SC_PAGE_SIZE);
 #endif
 #endif
